+++ release-notes.md
@@ -2,6 +2,12 @@
 
 ### Features:
 - Improve Options To Exit,Go to Home Or Play Again Fixes 
+- Add Pause Game Buttons
+- Cool Island Transparent Animation For Pause Game
+- Restarting The Game
+- 
 
 ### Bug Fixes:
  - Confetti Not Showing Or Making Dialog Span Whole Width Fixes #4
+ - Fix Game Play Starts immediately While Dialog Animates Away
+ -